requirements.txt
asgiref
certifi
distlib
Django
django-crispy-forms
django-email
django-environ
django-filter
djangorestframework
django-cors-headers
filelock
Markdown
pillow
pipenv
platformdirs
pytz
setuptools
sqlparse
tzdata
virtualenv
